Refactor tracers
