+++ src/pynguin/ga/computations.py
@@ -20,18 +20,18 @@ from typing import TypeVar
 
 import pynguin.utils.opcodes as op
 
+from pynguin.instrumentation.tracer import ExecutionTrace
 from pynguin.slicer.dynamicslicer import AssertionSlicer
 from pynguin.slicer.dynamicslicer import DynamicSlicer
-from pynguin.testcase.execution import ExecutionTrace
 
 
 if TYPE_CHECKING:
     from collections.abc import Callable
 
+    from pynguin.instrumentation.tracer import SubjectProperties
     from pynguin.slicer.dynamicslicer import SlicingCriterion
     from pynguin.testcase.execution import AbstractTestCaseExecutor
     from pynguin.testcase.execution import ExecutionResult
-    from pynguin.testcase.execution import SubjectProperties
     from pynguin.testcase.statement import Statement
