+++ src/pynguin/instrumentation/instrumentation.py
@@ -23,6 +23,7 @@ from bytecode import Bytecode
 from bytecode import ControlFlowGraph
 from bytecode import Instr
 
+import pynguin.instrumentation.tracer as tr
 import pynguin.utils.opcodes as op
 
 from pynguin.analyses.constants import DynamicConstantProvider
@@ -32,7 +33,6 @@ from pynguin.analyses.controlflow import ControlDependenceGraph
 
 if TYPE_CHECKING:
     from pynguin.analyses.controlflow import ProgramGraphNode
-    from pynguin.testcase.execution import ExecutionTracer
 
 CODE_OBJECT_ID_KEY = "code_object_id"
 
@@ -78,6 +78,22 @@ class CodeObjectMetaData:
 
     # CDG of this Code Object
     cdg: ControlDependenceGraph
+
+    def __getstate__(self) -> dict:
+        return {
+            "code_object": self.code_object,
+            "parent_code_object_id": self.parent_code_object_id,
+            "cfg": self.cfg,
+            "original_cfg": self.original_cfg,
+            "cdg": self.cdg,
+        }
+
+    def __setstate__(self, state: dict) -> None:
+        self.code_object = state["code_object"]
+        self.parent_code_object_id = state["parent_code_object_id"]
+        self.cfg = state["cfg"]
+        self.original_cfg = state["original_cfg"]
+        self.cdg = state["cdg"]
 
 
 @dataclass
@@ -218,11 +234,20 @@ class InstrumentationTransformer:
 
     def __init__(  # noqa: D107
         self,
-        tracer: ExecutionTracer,
+        instrumentation_tracer: tr.InstrumentationExecutionTracer,
         instrumentation_adapters: list[InstrumentationAdapter],
     ):
         self._instrumentation_adapters = instrumentation_adapters
-        self._tracer = tracer
+        self._instrumentation_tracer = instrumentation_tracer
+
+    @property
+    def instrumentation_tracer(self) -> tr.InstrumentationExecutionTracer:
+        """Get the tracer that is used for the instrumentation.
+
+        Returns:
+            The tracer that is used for the instrumentation.
+        """
+        return self._instrumentation_tracer
 
     def instrument_module(self, module_code: CodeType) -> CodeType:
         """Instrument the given code object of a module.
@@ -258,7 +283,7 @@ class InstrumentationTransformer:
         cfg = CFG.from_bytecode(Bytecode.from_code(code))
         original_cfg = CFG.from_bytecode(Bytecode.from_code(code))
         cdg = ControlDependenceGraph.compute(cfg)
-        code_object_id = self._tracer.register_code_object(
+        code_object_id = self._instrumentation_tracer.register_code_object(
             CodeObjectMetaData(
                 code_object=code,
                 parent_code_object_id=parent_code_object_id,
@@ -336,8 +361,10 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
 
     _logger = logging.getLogger(__name__)
 
-    def __init__(self, tracer: ExecutionTracer) -> None:  # noqa: D107
-        self._tracer = tracer
+    def __init__(  # noqa: D107
+        self, instrumentation_tracer: tr.InstrumentationExecutionTracer
+    ) -> None:
+        self._instrumentation_tracer = instrumentation_tracer
 
     def visit_node(
         self,
@@ -444,7 +471,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
             The id assigned to the predicate.
         """
         lineno = block[self._JUMP_OP_POS].lineno  # type: ignore[union-attr]
-        predicate_id = self._tracer.register_predicate(
+        predicate_id = self._instrumentation_tracer.register_predicate(
             PredicateMetaData(
                 line_no=lineno,  # type: ignore[arg-type]
                 code_object_id=code_object_id,
@@ -458,7 +485,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
             ArtificialInstr("DUP_TOP", lineno=lineno),
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -501,7 +528,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
             The id assigned to the predicate.
         """
         lineno = block[self._JUMP_OP_POS].lineno  # type: ignore[union-attr]
-        predicate_id = self._tracer.register_predicate(
+        predicate_id = self._instrumentation_tracer.register_predicate(
             PredicateMetaData(
                 line_no=lineno,  # type: ignore[arg-type]
                 code_object_id=code_object_id,
@@ -538,7 +565,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
             ArtificialInstr("DUP_TOP_TWO", lineno=lineno),
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -574,7 +601,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
             The id assigned to the predicate.
         """
         lineno = basic_block[self._JUMP_OP_POS].lineno  # type: ignore[union-attr]
-        predicate_id = self._tracer.register_predicate(
+        predicate_id = self._instrumentation_tracer.register_predicate(
             PredicateMetaData(
                 line_no=lineno,  # type: ignore[arg-type]
                 code_object_id=code_object_id,
@@ -588,7 +615,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
             ArtificialInstr("DUP_TOP_TWO", lineno=lineno),
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -624,7 +651,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
         basic_block[0:0] = [
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -683,7 +710,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
         for_instr = basic_block[self._JUMP_OP_POS]
         assert for_instr.opcode == op.FOR_ITER  # type: ignore[union-attr]
         lineno = for_instr.lineno  # type: ignore[union-attr]
-        predicate_id = self._tracer.register_predicate(
+        predicate_id = self._instrumentation_tracer.register_predicate(
             PredicateMetaData(
                 line_no=lineno,  # type: ignore[arg-type]
                 code_object_id=code_object_id,
@@ -701,7 +728,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
         entered.extend([
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -725,7 +752,7 @@ class BranchCoverageInstrumentation(InstrumentationAdapter):
         not_entered.extend([
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -753,8 +780,10 @@ class LineCoverageInstrumentation(InstrumentationAdapter):
 
     _logger = logging.getLogger(__name__)
 
-    def __init__(self, tracer: ExecutionTracer) -> None:  # noqa: D107
-        self._tracer = tracer
+    def __init__(  # noqa: D107
+        self, instrumentation_tracer: tr.InstrumentationExecutionTracer
+    ) -> None:
+        self._instrumentation_tracer = instrumentation_tracer
 
     def visit_node(  # noqa: D102
         self,
@@ -771,7 +800,7 @@ class LineCoverageInstrumentation(InstrumentationAdapter):
         while instr_index < len(basic_block):
             if basic_block[instr_index].lineno != lineno:  # type: ignore[union-attr]
                 lineno = basic_block[instr_index].lineno  # type: ignore[union-attr]
-                line_id = self._tracer.register_line(
+                line_id = self._instrumentation_tracer.register_line(
                     code_object_id,
                     file_name,
                     lineno,  # type: ignore[arg-type]
@@ -805,7 +834,7 @@ class LineCoverageInstrumentation(InstrumentationAdapter):
         inserted_instructions = [
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=lineno,
             ),
             ArtificialInstr(
@@ -834,8 +863,10 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
 
     _logger = logging.getLogger(__name__)
 
-    def __init__(self, tracer: ExecutionTracer) -> None:  # noqa: D107
-        self._tracer = tracer
+    def __init__(  # noqa: D107
+        self, instrumentation_tracer: tr.InstrumentationExecutionTracer
+    ) -> None:
+        self._instrumentation_tracer = instrumentation_tracer
 
     def visit_node(  # noqa: C901
         self,
@@ -879,7 +910,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
                 and file_name != "<ast>"
             ):
                 lineno = instr.lineno  # type: ignore[union-attr]
-                self._tracer.register_line(
+                self._instrumentation_tracer.register_line(
                     code_object_id,
                     file_name,
                     lineno,  # type: ignore[arg-type]
@@ -1012,7 +1043,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1063,7 +1094,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1128,7 +1159,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1158,7 +1189,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             #   Load lookup method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer.__class__,  # type: ignore[arg-type]
+                self._instrumentation_tracer.__class__,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1239,7 +1270,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1302,7 +1333,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1353,7 +1384,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1405,7 +1436,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1464,7 +1495,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
@@ -1521,7 +1552,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             # references method in the ExecutionTracer by name
@@ -1566,7 +1597,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             # references method in the ExecutionTracer by name
@@ -1600,7 +1631,7 @@ class CheckedCoverageInstrumentation(InstrumentationAdapter):
             # Load tracing method
             ArtificialInstr(
                 "LOAD_CONST",
-                self._tracer,  # type: ignore[arg-type]
+                self._instrumentation_tracer,  # type: ignore[arg-type]
                 lineno=instr.lineno,
             ),
             ArtificialInstr(
