+++ src/pynguin/testcase/execution.py
@@ -20,17 +20,9 @@ import sys
 import threading
 
 from abc import abstractmethod
-from collections.abc import Sized
-from dataclasses import dataclass
-from dataclasses import field
 from importlib import reload
-from math import inf
 from queue import Empty
 from queue import Queue
-from types import BuiltinFunctionType
-from types import BuiltinMethodType
-from types import MappingProxyType
-from types import ModuleType
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import TypeVar
@@ -39,15 +31,10 @@ from typing import cast
 # Needs to be loaded, i.e., in sys.modules for the execution of assertions to work.
 import pytest  # noqa: F401
 
-from bytecode import BasicBlock
-from bytecode import CellVar
-from bytecode import FreeVar
-
 import pynguin.assertion.assertion as ass
 import pynguin.assertion.assertion_to_ast as ass_to_ast
 import pynguin.assertion.assertion_trace as at
 import pynguin.configuration as config
-import pynguin.slicer.executedinstruction as ei
 import pynguin.testcase.statement as stmt
 import pynguin.testcase.statement_to_ast as stmt_to_ast
 import pynguin.testcase.variablereference as vr
@@ -62,25 +49,22 @@ from pynguin.analyses.typesystem import ProperType
 from pynguin.analyses.typesystem import TupleType
 from pynguin.instrumentation.instrumentation import ArtificialInstr
 from pynguin.instrumentation.instrumentation import CheckedCoverageInstrumentation
-from pynguin.instrumentation.instrumentation import CodeObjectMetaData
 from pynguin.instrumentation.instrumentation import InstrumentationTransformer
-from pynguin.instrumentation.instrumentation import PredicateMetaData
-from pynguin.instrumentation.instrumentation import PynguinCompare
+from pynguin.instrumentation.tracer import ExecutionTrace
+from pynguin.instrumentation.tracer import ExecutionTracer
+from pynguin.instrumentation.tracer import InstrumentationExecutionTracer
 from pynguin.testcase.mocking import mocks_to_use
 from pynguin.utils.mirror import Mirror
-from pynguin.utils.orderedset import OrderedSet
-from pynguin.utils.type_utils import given_exception_matches
-from pynguin.utils.type_utils import is_bytes
-from pynguin.utils.type_utils import is_numeric
-from pynguin.utils.type_utils import is_string
-from pynguin.utils.type_utils import string_distance
-from pynguin.utils.type_utils import string_le_distance
-from pynguin.utils.type_utils import string_lt_distance
 
 
 immutable_types = (int, float, complex, str, tuple, frozenset, bytes)
 
 if TYPE_CHECKING:
+    from types import MappingProxyType
+    from types import ModuleType
+
+    from bytecode import BasicBlock
+
     import pynguin.testcase.testcase as tc
 
     from pynguin.analyses import module
@@ -554,258 +538,6 @@ class ReturnTypeObserver(ExecutionObserver):
                 )
 
 
-@dataclass
-class ExecutionTrace:
-    """Stores trace information about the execution."""
-
-    _logger = logging.getLogger(__name__)
-
-    executed_code_objects: OrderedSet[int] = field(default_factory=OrderedSet)
-    executed_predicates: dict[int, int] = field(default_factory=dict)
-    true_distances: dict[int, float] = field(default_factory=dict)
-    false_distances: dict[int, float] = field(default_factory=dict)
-    covered_line_ids: OrderedSet[int] = field(default_factory=OrderedSet)
-    executed_instructions: list[ei.ExecutedInstruction] = field(default_factory=list)
-    executed_assertions: list[ExecutedAssertion] = field(default_factory=list)
-    checked_lines: OrderedSet[int] = field(default_factory=OrderedSet)
-
-    def merge(self, other: ExecutionTrace) -> None:
-        """Merge the values from the other execution trace.
-
-        Args:
-            other: Merges the other traces into this trace
-        """
-        self.executed_code_objects.update(other.executed_code_objects)
-        for key, value in other.executed_predicates.items():
-            self.executed_predicates[key] = self.executed_predicates.get(key, 0) + value
-        self._merge_min(self.true_distances, other.true_distances)
-        self._merge_min(self.false_distances, other.false_distances)
-        self.covered_line_ids.update(other.covered_line_ids)
-        self.checked_lines.update(other.checked_lines)
-        shift: int = len(self.executed_instructions)
-        self.executed_instructions.extend(other.executed_instructions)
-        self.executed_assertions.extend(
-            ExecutedAssertion(
-                executed_assertion.code_object_id,
-                executed_assertion.node_id,
-                executed_assertion.trace_position + shift,
-                executed_assertion.assertion,
-            )
-            for executed_assertion in other.executed_assertions
-        )
-
-    @staticmethod
-    def _merge_min(target: dict[int, float], source: dict[int, float]) -> None:
-        """Merge source into target. Minimum value wins.
-
-        Args:
-            target: the target to merge the values in
-            source: the source of the merge
-        """
-        for key, value in source.items():
-            target[key] = min(target.get(key, inf), value)
-
-    def update_predicate_distances(
-        self, distance_true: float, distance_false: float, predicate: int
-    ) -> None:
-        """Update the distances and predicate execution count.
-
-        Args:
-            distance_true: the measured true distance
-            distance_false: the measured false distance
-            predicate: the predicate id
-        """
-        self.executed_predicates[predicate] = self.executed_predicates.get(predicate, 0) + 1
-        self.true_distances[predicate] = min(self.true_distances.get(predicate, inf), distance_true)
-        self.false_distances[predicate] = min(
-            self.false_distances.get(predicate, inf), distance_false
-        )
-
-    def add_instruction(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-    ) -> None:
-        """Creates a new ExecutedInstruction object and adds it to the trace.
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-        """
-        executed_instr = ei.ExecutedInstruction(
-            module, code_object_id, node_id, opcode, None, lineno, offset
-        )
-        self.executed_instructions.append(executed_instr)
-
-    def add_memory_instruction(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        arg_name: str,
-        arg_address: int,
-        is_mutable_type: bool,  # noqa: FBT001
-        object_creation: bool,  # noqa: FBT001
-    ) -> None:
-        """Creates a new ExecutedMemoryInstruction object and adds it to the trace.
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            arg_name: the name of the argument
-            arg_address: the memory address of the argument
-            is_mutable_type: if the argument is mutable
-            object_creation: if the instruction creates the object used
-        """
-        executed_instr = ei.ExecutedMemoryInstruction(
-            module,
-            code_object_id,
-            node_id,
-            opcode,
-            arg_name,
-            lineno,
-            offset,
-            arg_address,
-            is_mutable_type,
-            object_creation,
-        )
-        self.executed_instructions.append(executed_instr)
-
-    def add_attribute_instruction(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        attr_name: str,
-        src_address: int,
-        arg_address: int,
-        is_mutable_type: bool,  # noqa: FBT001
-    ) -> None:
-        """Creates a new ExecutedAttributeInstruction object and adds it to the trace.
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            attr_name: the name of the accessed attribute
-            src_address: the memory address of the attribute
-            arg_address: the memory address of the argument
-            is_mutable_type: if the attribute is mutable
-        """
-        executed_instr = ei.ExecutedAttributeInstruction(
-            module,
-            code_object_id,
-            node_id,
-            opcode,
-            attr_name,
-            lineno,
-            offset,
-            src_address,
-            arg_address,
-            is_mutable_type,
-        )
-        self.executed_instructions.append(executed_instr)
-
-    def add_jump_instruction(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        target_id: int,
-    ) -> None:
-        """Creates a new ExecutedControlInstruction object and adds it to the trace.
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            target_id: the target offset to jump to
-        """
-        executed_instr = ei.ExecutedControlInstruction(
-            module, code_object_id, node_id, opcode, target_id, lineno, offset
-        )
-        self.executed_instructions.append(executed_instr)
-
-    def add_call_instruction(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        arg: int,
-    ) -> None:
-        """Creates a new ExecutedCallInstruction object and adds it to the trace.
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            arg: the argument to the instruction
-        """
-        executed_instr = ei.ExecutedCallInstruction(
-            module, code_object_id, node_id, opcode, arg, lineno, offset
-        )
-
-        self.executed_instructions.append(executed_instr)
-
-    def add_return_instruction(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-    ) -> None:
-        """Creates a new ExecutedReturnInstruction object and adds it to the trace.
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-        """
-        executed_instr = ei.ExecutedReturnInstruction(
-            module, code_object_id, node_id, opcode, None, lineno, offset
-        )
-
-        self.executed_instructions.append(executed_instr)
-
-
 @dataclasses.dataclass
 class ExecutionResult:
     """Result of an execution."""
@@ -917,1023 +649,6 @@ class ExecutionResult:
 
     def __repr__(self) -> str:
         return str(self)
-
-
-@dataclass
-class LineMetaData:
-    """Stores meta data of a line."""
-
-    # id of the code object where the line is first defined
-    code_object_id: int
-
-    # name of the file containing a line
-    file_name: str
-
-    # Line number where the predicate is defined.
-    line_number: int
-
-    def __hash__(self):
-        # code object id is not checked since file
-        # and line number are the unique identifiers
-        return 31 + self.line_number + hash(self.file_name)
-
-    def __eq__(self, other: object) -> bool:
-        if self is other:
-            return True
-        if not isinstance(other, LineMetaData):
-            return False
-        # code object id is not checked since file
-        # and line number are the unique identifiers
-        return self.line_number == other.line_number and self.file_name == other.file_name
-
-
-@dataclass
-class SubjectProperties:
-    """Contains properties about the subject under test."""
-
-    # Maps all known ids of Code Objects to meta information
-    existing_code_objects: dict[int, CodeObjectMetaData] = field(default_factory=dict)
-
-    # Stores which of the existing code objects do not contain a branch, i.e.,
-    # they do not contain a predicate. Every code object is initially seen as
-    # branch-less until a predicate is registered for it.
-    branch_less_code_objects: OrderedSet[int] = field(default_factory=OrderedSet)
-
-    # Maps all known ids of predicates to meta information
-    existing_predicates: dict[int, PredicateMetaData] = field(default_factory=dict)
-
-    # stores which line id represents which line in which file
-    existing_lines: dict[int, LineMetaData] = field(default_factory=dict)
-
-    # stores known memory attribute object addresses
-    object_addresses: OrderedSet[int] = field(default_factory=OrderedSet)
-
-
-class ExecutionTracer:  # noqa: PLR0904
-    """Tracks branch distances and covered statements during execution.
-
-    The results are stored in an execution trace.
-    """
-
-    _logger = logging.getLogger(__name__)
-
-    class TracerLocalState(threading.local):
-        """Encapsulate state that is thread specific."""
-
-        def __init__(self):  # noqa: D107
-            super().__init__()
-            self.enabled = True
-            self.trace = ExecutionTrace()
-
-    def __init__(self) -> None:  # noqa: D107
-        self.subject_properties = SubjectProperties()
-        # Contains the trace information that is generated when a module is imported
-        self._import_trace = ExecutionTrace()
-
-        # Thread local state
-        self._thread_local_state = ExecutionTracer.TracerLocalState()
-
-        self.init_trace()
-        self._current_thread_identifier: int | None = None
-
-    @property
-    def current_thread_identifier(self) -> int | None:
-        """Get the current thread identifier.
-
-        Returns:
-            The current thread identifier
-        """
-        return self._current_thread_identifier
-
-    @current_thread_identifier.setter
-    def current_thread_identifier(self, current: int) -> None:
-        """Set the current thread identifier.
-
-        Tracing calls from any other thread are ignored.
-
-        Args:
-            current: the current thread
-        """
-        self._current_thread_identifier = current
-
-    @property
-    def import_trace(self) -> ExecutionTrace:
-        """The trace that was generated when the SUT was imported.
-
-        Returns:
-            The execution trace after executing the import statements
-        """
-        copied = ExecutionTrace()
-        copied.merge(self._import_trace)
-        return copied
-
-    def get_subject_properties(self) -> SubjectProperties:
-        """Provide known data.
-
-        Returns:
-            The known data about the execution
-        """
-        return self.subject_properties
-
-    def reset(self) -> None:
-        """Resets everything.
-
-        Should be called before instrumentation. Clears all data, so we can handle a
-        reload of the SUT.
-        """
-        self.subject_properties = SubjectProperties()
-        self._import_trace = ExecutionTrace()
-        self.init_trace()
-
-    def store_import_trace(self) -> None:
-        """Stores the current trace as the import trace.
-
-        Should only be done once, after a module was loaded. The import trace will be
-        merged into every subsequently recorded trace.
-        """
-        self._import_trace = self._thread_local_state.trace
-        self.init_trace()
-
-    def init_trace(self) -> None:
-        """Create a new trace that only contains the trace data from the import."""
-        new_trace = ExecutionTrace()
-        new_trace.merge(self._import_trace)
-        self._thread_local_state.trace = new_trace
-
-    def is_disabled(self) -> bool:
-        """Should we track anything?
-
-        We might have to disable tracing, e.g. when calling __eq__ ourselves.
-        Otherwise, we create an endless recursion.
-
-        Returns:
-            Whether we should track anything
-        """
-        return not self._thread_local_state.enabled
-
-    def enable(self) -> None:
-        """Enable tracing."""
-        self._thread_local_state.enabled = True
-
-    def disable(self) -> None:
-        """Disable tracing."""
-        self._thread_local_state.enabled = False
-
-    def get_trace(self) -> ExecutionTrace:
-        """Get the trace with the current information.
-
-        Returns:
-            The current execution trace
-        """
-        return self._thread_local_state.trace
-
-    def register_code_object(self, meta: CodeObjectMetaData) -> int:
-        """Declare that a code object exists.
-
-        Args:
-            meta: the code objects existing
-
-        Returns:
-            the id of the code object, which can be used to identify the object
-            during instrumentation.
-        """
-        code_object_id = len(self.subject_properties.existing_code_objects)
-        self.subject_properties.existing_code_objects[code_object_id] = meta
-        self.subject_properties.branch_less_code_objects.add(code_object_id)
-        return code_object_id
-
-    def executed_code_object(self, code_object_id: int) -> None:
-        """Mark a code object as executed.
-
-        This means, that the routine which refers to this code object was at least
-        called once.
-
-        Args:
-            code_object_id: the code object id to mark
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        assert code_object_id in self.subject_properties.existing_code_objects, (
-            "Cannot trace unknown code object"
-        )
-        self._thread_local_state.trace.executed_code_objects.add(code_object_id)
-
-    def register_predicate(self, meta: PredicateMetaData) -> int:
-        """Declare that a predicate exists.
-
-        Args:
-            meta: Metadata about the predicates
-
-        Returns:
-            the id of the predicate, which can be used to identify the predicate
-            during instrumentation.
-        """
-        predicate_id = len(self.subject_properties.existing_predicates)
-        self.subject_properties.existing_predicates[predicate_id] = meta
-        self.subject_properties.branch_less_code_objects.discard(meta.code_object_id)
-        return predicate_id
-
-    def executed_compare_predicate(  # noqa: C901
-        self, value1, value2, predicate: int, cmp_op: PynguinCompare
-    ) -> None:
-        """A predicate that is based on a comparison was executed.
-
-        Args:
-            value1: the first value
-            value2: the second value
-            predicate: the predicate identifier
-            cmp_op: the compare operation
-
-        Raises:
-            RuntimeError: raised when called from another thread.
-            AssertionError: when encountering an unknown compare op.
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        try:
-            self.disable()
-            assert predicate in self.subject_properties.existing_predicates, (
-                "Cannot trace unknown predicate"
-            )
-            value1 = tt.unwrap(value1)
-            value2 = tt.unwrap(value2)
-
-            match cmp_op:
-                case PynguinCompare.EQ:
-                    distance_true, distance_false = (
-                        _eq(value1, value2),
-                        _neq(value1, value2),
-                    )
-                case PynguinCompare.NE:
-                    distance_true, distance_false = (
-                        _neq(value1, value2),
-                        _eq(value1, value2),
-                    )
-                case PynguinCompare.LT:
-                    distance_true, distance_false = (
-                        _lt(value1, value2),
-                        _le(value2, value1),
-                    )
-                case PynguinCompare.LE:
-                    distance_true, distance_false = (
-                        _le(value1, value2),
-                        _lt(value2, value1),
-                    )
-                case PynguinCompare.GT:
-                    distance_true, distance_false = (
-                        _lt(value2, value1),
-                        _le(value1, value2),
-                    )
-                case PynguinCompare.GE:
-                    distance_true, distance_false = (
-                        _le(value2, value1),
-                        _lt(value1, value2),
-                    )
-                case PynguinCompare.IN:
-                    distance_true, distance_false = (
-                        _in(value1, value2),
-                        _nin(value1, value2),
-                    )
-                case PynguinCompare.NOT_IN:
-                    distance_true, distance_false = (
-                        _nin(value1, value2),
-                        _in(value1, value2),
-                    )
-                case PynguinCompare.IS:
-                    distance_true, distance_false = (
-                        _is(value1, value2),
-                        _isn(value1, value2),
-                    )
-                case PynguinCompare.IS_NOT:
-                    distance_true, distance_false = (
-                        _isn(value1, value2),
-                        _is(value1, value2),
-                    )
-                case _:
-                    raise AssertionError("Unknown compare op")
-            self._update_metrics(distance_false, distance_true, predicate)
-        finally:
-            self.enable()
-
-    def executed_bool_predicate(self, value, predicate: int) -> None:
-        """A predicate that is based on a boolean value was executed.
-
-        Args:
-            value: the value
-            predicate: the predicate identifier
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        try:
-            self.disable()
-            assert predicate in self.subject_properties.existing_predicates, (
-                "Cannot trace unknown predicate"
-            )
-            distance_true = 0.0
-            distance_false = 0.0
-            # Might be necessary when using Proxies.
-            value = tt.unwrap(value)
-            if value:
-                if isinstance(value, Sized):
-                    # Sized instances evaluate to False if they are empty,
-                    # and to True otherwise, thus we can use their size as a distance
-                    # measurement.
-                    distance_false = len(value)
-                elif is_numeric(value):
-                    # For numeric value, we can use their absolute value
-                    distance_false = float(abs(value))
-                else:
-                    # Necessary to use inf instead of 1.0 here,
-                    # so that a value for which we can't compute a false distance
-                    # always has the greatest distance to the false branch than an
-                    # object for which we can compute a distance.
-                    distance_false = inf
-            else:
-                distance_true = 1.0
-
-            self._update_metrics(distance_false, distance_true, predicate)
-        finally:
-            self.enable()
-
-    def executed_exception_match(self, err, exc, predicate: int):
-        """A predicate that is based on exception matching was executed.
-
-        Args:
-            err: The raised exception
-            exc: The matching condition
-            predicate: the predicate identifier
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        try:
-            self.disable()
-            assert predicate in self.subject_properties.existing_predicates, (
-                "Cannot trace unknown predicate"
-            )
-            distance_true = 0.0
-            distance_false = 0.0
-            # Might be necessary when using Proxies.
-            err = tt.unwrap(err)
-            exc = tt.unwrap(exc)
-            if given_exception_matches(err, exc):
-                distance_false = 1.0
-            else:
-                distance_true = 1.0
-
-            self._update_metrics(distance_false, distance_true, predicate)
-        finally:
-            self.enable()
-
-    def track_line_visit(self, line_id: int) -> None:
-        """Tracks the visit of a line.
-
-        Args:
-            line_id: the if of the line that was visited
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        self._thread_local_state.trace.covered_line_ids.add(line_id)
-
-    def register_line(self, code_object_id: int, file_name: str, line_number: int) -> int:
-        """Tracks the existence of a line.
-
-        Args:
-            code_object_id: The id of the code object that contains the line
-            file_name: The file in which the statement is
-            line_number: The line of the statement to track
-
-        Returns:
-            the id of the registered line
-        """
-        line_meta = LineMetaData(code_object_id, file_name, line_number)
-        if line_meta not in self.subject_properties.existing_lines.values():
-            line_id = len(self.subject_properties.existing_lines)
-            self.subject_properties.existing_lines[line_id] = line_meta
-        else:
-            index = list(self.subject_properties.existing_lines.values()).index(line_meta)
-            line_id = list(self.subject_properties.existing_lines.keys())[index]
-        return line_id
-
-    def _update_metrics(self, distance_false: float, distance_true: float, predicate: int):
-        assert predicate in self.subject_properties.existing_predicates, (
-            "Cannot update unknown predicate"
-        )
-        assert distance_true >= 0.0, "True distance cannot be negative"
-        assert distance_false >= 0.0, "False distance cannot be negative"
-        assert (distance_true == 0.0) ^ (distance_false == 0.0), (
-            "Exactly one distance must be 0.0, i.e., one branch must be taken."
-        )
-        self._thread_local_state.trace.update_predicate_distances(
-            distance_true=distance_true,
-            distance_false=distance_false,
-            predicate=predicate,
-        )
-
-    def track_generic(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-    ) -> None:
-        """Track a generic instruction inside the trace.
-
-        Note: This method is referenced by name in the instrumentation
-        for checked coverage. To avoid circular imports, the name is simply written
-        as string, so if this method is renamed, please adjust the string in the
-        instrumentation. Otherwise, the checked coverage instrumentation breaks!
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        self._thread_local_state.trace.add_instruction(
-            module, code_object_id, node_id, opcode, lineno, offset
-        )
-
-    def track_memory_access(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        arg: str | CellVar | FreeVar,
-        arg_address: int,
-        arg_type: type,
-    ) -> None:
-        """Track a memory access instruction in the trace.
-
-        Note: This method is referenced by name in the instrumentation
-        for checked coverage. To avoid circular imports, the name is simply written
-        as string, so if this method is renamed, please adjust the string in the
-        instrumentation. Otherwise, the checked coverage instrumentation breaks!
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            arg: the used variable
-            arg_address: the memory address of the variable
-            arg_type: the type of the variable
-
-        Raises:
-            ValueError: when no argument is given
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        if not arg and opcode != op.IMPORT_NAME:  # IMPORT_NAMEs may not have arguments
-            raise ValueError("A memory access instruction must have an argument")
-        if isinstance(arg, CellVar | FreeVar):
-            arg = arg.name
-
-        # Determine if this is a mutable type
-        mutable_type = True
-        if arg_type in immutable_types:
-            mutable_type = False
-
-        # Determine if this is a definition of a completely new object
-        # (required later during slicing)
-        object_creation = False
-        if arg_address and arg_address not in self.get_subject_properties().object_addresses:
-            object_creation = True
-            self.subject_properties.object_addresses.add(arg_address)
-
-        self._thread_local_state.trace.add_memory_instruction(
-            module,
-            code_object_id,
-            node_id,
-            opcode,
-            lineno,
-            offset,
-            arg,
-            arg_address,
-            mutable_type,
-            object_creation,
-        )
-
-    def track_attribute_access(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        attr_name: str,
-        src_address: int,
-        arg_address: int,
-        arg_type: type,
-    ) -> None:
-        """Track an attribute access instruction in the trace.
-
-        Note: This method is referenced by name in the instrumentation
-        for checked coverage. To avoid circular imports, the name is simply written
-        as string, so if this method is renamed, please adjust the string in the
-        instrumentation. Otherwise, the checked coverage instrumentation breaks!
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            attr_name: the name of the accessed attribute
-            src_address: the memory address of the attribute
-            arg_address: the memory address of the argument
-            arg_type: the type of argument
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        # Different built-in methods and functions often have the same address when
-        # accessed sequentially.
-        # The address is not recorded in such cases.
-        if arg_type is BuiltinMethodType or arg_type is BuiltinFunctionType:
-            arg_address = -1
-
-        # Determine if this is a mutable type
-        mutable_type = True
-        if arg_type in immutable_types:
-            mutable_type = False
-
-        self._thread_local_state.trace.add_attribute_instruction(
-            module,
-            code_object_id,
-            node_id,
-            opcode,
-            lineno,
-            offset,
-            attr_name,
-            src_address,
-            arg_address,
-            mutable_type,
-        )
-
-    def track_jump(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        target_id: int,
-    ) -> None:
-        """Track a jump instruction in the trace.
-
-        Note: This method is referenced by name in the instrumentation
-        for checked coverage. To avoid circular imports, the name is simply written
-        as string, so if this method is renamed, please adjust the string in the
-        instrumentation. Otherwise, the checked coverage instrumentation breaks!
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            target_id: the offset of the target of the jump
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        self._thread_local_state.trace.add_jump_instruction(
-            module, code_object_id, node_id, opcode, lineno, offset, target_id
-        )
-
-    def track_call(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-        arg: int,
-    ) -> None:
-        """Track a method call instruction in the trace.
-
-        Note: This method is referenced by name in the instrumentation
-        for checked coverage. To avoid circular imports, the name is simply written
-        as string, so if this method is renamed, please adjust the string in the
-        instrumentation. Otherwise, the checked coverage instrumentation breaks!
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-            arg: the argument used in the method call
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        self._thread_local_state.trace.add_call_instruction(
-            module, code_object_id, node_id, opcode, lineno, offset, arg
-        )
-
-    def track_return(  # noqa: PLR0917
-        self,
-        module: str,
-        code_object_id: int,
-        node_id: int,
-        opcode: int,
-        lineno: int,
-        offset: int,
-    ) -> None:
-        """Track a return instruction in the trace.
-
-        Note: This method is referenced by name in the instrumentation
-        for checked coverage. To avoid circular imports, the name is simply written
-        as string, so if this method is renamed, please adjust the string in the
-        instrumentation. Otherwise, the checked coverage instrumentation breaks!
-
-        Args:
-            module: File name of the module containing the instruction
-            code_object_id: code object containing the instruction
-            node_id: the node of the code object containing the instruction
-            opcode: the opcode of the instruction
-            lineno: the line number of the instruction
-            offset: the offset of the instruction
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        self._thread_local_state.trace.add_return_instruction(
-            module, code_object_id, node_id, opcode, lineno, offset
-        )
-
-    def register_exception_assertion(self, statement: stmt.Statement) -> None:
-        """Track the position of an exception assertion in the trace.
-
-        Normally, to track an assertion, we trace the POP_JUMP_IF_TRUE instruction
-        contained by each assertion. The pytest exception assertion does not use
-        an assertion containing this instruction.
-        Therefore, we trace the instruction that was last executed before
-        the exception.
-
-        Args:
-            statement: the statement causing the exception
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        if statement.has_only_exception_assertion():
-            trace = self._thread_local_state.trace
-            error_call_position = len(trace.executed_instructions) - 1
-            error_causing_instr = trace.executed_instructions[error_call_position]
-            code_object_id = error_causing_instr.code_object_id
-            node_id = error_causing_instr.node_id
-            trace.executed_assertions.append(
-                ExecutedAssertion(
-                    code_object_id,
-                    node_id,
-                    error_call_position,
-                    next(iter(statement.assertions)),
-                )
-            )
-
-    def register_assertion_position(
-        self, code_object_id: int, node_id: int, assertion: ass.Assertion
-    ) -> None:
-        """Track the position of an assertion in the trace.
-
-        Args:
-            code_object_id: code object containing the assertion to register
-            node_id: the id of the node containing the assertion to register
-            assertion: the assertion of the statement
-
-        Raises:
-            RuntimeError: raised when called from another thread
-        """
-        if threading.current_thread().ident != self._current_thread_identifier:
-            raise RuntimeError("The current thread shall not be executed any more, thus I kill it.")
-
-        if self.is_disabled():
-            return
-
-        exec_instr = self.get_trace().executed_instructions
-        pop_jump_if_true_position = len(exec_instr) - 1
-        for instr in reversed(exec_instr):
-            if instr.opcode == op.POP_JUMP_IF_TRUE:
-                break
-            pop_jump_if_true_position -= 1
-        assert pop_jump_if_true_position != -1, (
-            "Node in code object did not contain a POP_JUMP_IF_TRUE instruction"
-        )
-
-        self._thread_local_state.trace.executed_assertions.append(
-            ExecutedAssertion(
-                code_object_id,
-                node_id,
-                pop_jump_if_true_position,
-                assertion,
-            )
-        )
-
-    @staticmethod
-    def attribute_lookup(object_type, attribute: str) -> int:
-        """Check the dictionary of classes making up the MRO (_PyType_Lookup).
-
-        The attribute must be a data descriptor to be prioritized here
-
-        Args:
-            object_type: The type object to check
-            attribute: the attribute to check for in the class
-
-        Returns:
-            The id of the object type or the class if it has the attribute, -1 otherwise
-        """
-        for cls in type(object_type).__mro__:
-            if attribute in cls.__dict__ and inspect.isdatadescriptor(cls.__dict__.get(attribute)):
-                # Class in the MRO hierarchy has attribute
-                # Class has attribute and attribute is a data descriptor
-                return id(cls)
-
-        # This would lead to an infinite recursion and thus a crash of the program
-        if attribute in {"__getattr__", "__getitem__"}:
-            return -1
-        # Check if the dictionary of the object on which lookup is performed
-        if (
-            hasattr(object_type, "__dict__")
-            and object_type.__dict__
-            and attribute in object_type.__dict__
-        ):
-            return id(object_type)
-        if (
-            hasattr(object_type, "__slots__")
-            and object_type.__slots__
-            and attribute in object_type.__slots__
-        ):
-            return id(object_type)
-
-        # Check if attribute in MRO hierarchy (no need for data descriptor)
-        for cls in type(object_type).__mro__:
-            if attribute in cls.__dict__:
-                return id(cls)
-
-        return -1
-
-    def __repr__(self) -> str:
-        return "ExecutionTracer"
-
-    def lineids_to_linenos(self, line_ids: OrderedSet[int]) -> OrderedSet[int]:
-        """Convenience method to translate line ids to line numbers.
-
-        Args:
-            line_ids: The ids that should be translated.
-
-        Returns:
-            The line numbers.
-        """
-        return OrderedSet([
-            self.subject_properties.existing_lines[line_id].line_number for line_id in line_ids
-        ])
-
-
-@dataclass
-class ExecutedAssertion:
-    """Data class for assertions of a testcase traced during execution for slicing."""
-
-    # the code object containing the executed assertion
-    code_object_id: int
-    # the node containing the executed assertion
-    node_id: int
-    # the position inside the exection trace of the executed assertion
-    trace_position: int
-    # the assertion object of a statement that was executed
-    assertion: ass.Assertion
-
-
-def _eq(val1, val2) -> float:
-    """Distance computation for '=='.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 == val2:
-        return 0.0
-    if is_numeric(val1) and is_numeric(val2):
-        return float(abs(val1 - val2))
-    if is_string(val1) and is_string(val2):
-        return string_distance(val1, val2)
-    if is_bytes(val1) and is_bytes(val2):
-        return string_distance(val1.decode("iso-8859-1"), val2.decode("iso-8859-1"))
-    return inf
-
-
-def _neq(val1, val2) -> float:
-    """Distance computation for '!='.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 != val2:
-        return 0.0
-    return 1.0
-
-
-def _lt(val1, val2) -> float:
-    """Distance computation for '<'.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 < val2:
-        return 0.0
-    if is_numeric(val1) and is_numeric(val2):
-        return (float(val1) - float(val2)) + 1.0
-    if is_string(val1) and is_string(val2):
-        return string_lt_distance(val1, val2)
-    if is_bytes(val1) and is_bytes(val2):
-        return string_lt_distance(val1.decode("iso-8859-1"), val2.decode("iso-8859-1"))
-    return inf
-
-
-def _le(val1, val2) -> float:
-    """Distance computation for '<='.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 <= val2:
-        return 0.0
-    if is_numeric(val1) and is_numeric(val2):
-        return float(val1) - float(val2)
-    if is_string(val1) and is_string(val2):
-        return string_le_distance(val1, val2)
-    if is_bytes(val1) and is_bytes(val2):
-        return string_le_distance(val1.decode("iso-8859-1"), val2.decode("iso-8859-1"))
-    return inf
-
-
-def _in(val1, val2) -> float:
-    """Distance computation for 'in'.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 in val2:
-        return 0.0
-    # TODO(fk) maybe limit this to certain collections?
-    #  Check only if collection size is within some range,
-    #  otherwise the check might take very long.
-
-    # Use the shortest distance to any element.
-    return min([_eq(val1, v) for v in val2] + [inf])
-
-
-def _nin(val1, val2) -> float:
-    """Distance computation for 'not in'.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 not in val2:
-        return 0.0
-    return 1.0
-
-
-def _is(val1, val2) -> float:
-    """Distance computation for 'is'.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 is val2:
-        return 0.0
-    return 1.0
-
-
-def _isn(val1, val2) -> float:
-    """Distance computation for 'is not'.
-
-    Args:
-        val1: the first value
-        val2: the second value
-
-    Returns:
-        the distance
-    """
-    if val1 is not val2:
-        return 0.0
-    return 1.0
 
 
 class ModuleProvider:
@@ -2113,16 +828,17 @@ class TestCaseExecutor(AbstractTestCaseExecutor):
         self._instrument = (
             config.CoverageMetric.CHECKED in config.configuration.statistics_output.coverage_metrics
         )
-        checked_instrumentation = CheckedCoverageInstrumentation(self._tracer)
+        instrumentation_tracer = InstrumentationExecutionTracer(self._tracer)
+        checked_instrumentation = CheckedCoverageInstrumentation(instrumentation_tracer)
         self._checked_transformer = InstrumentationTransformer(
-            self._tracer, [checked_instrumentation]
+            instrumentation_tracer, [checked_instrumentation]
         )
 
         def log_thread_exception(arg):
             _LOGGER.error(
                 "Exception in Thread: %s",
                 arg.thread,
-                exc_info=(arg.exc_type, arg.exc_value, arg.exc_traceback),  # noqa: LOG014
+                exc_info=(arg.exc_type, arg.exc_value, arg.exc_traceback),
             )
 
         # Set our own exception hook, so timeout related errors in executing threads
