+++ tests/ga/algorithms/test_integration_varioussearchalgorithms.py
@@ -18,7 +18,7 @@ import pynguin.ga.generationalgorithmfactory as gaf
 
 from pynguin.analyses.module import generate_test_cluster
 from pynguin.instrumentation.machinery import install_import_hook
-from pynguin.testcase.execution import ExecutionTracer
+from pynguin.instrumentation.tracer import ExecutionTracer
 from pynguin.testcase.execution import TestCaseExecutor
