+++ tests/ga/algorithms/test_mosaalgorithm.py
@@ -14,7 +14,7 @@ from pynguin.ga.algorithms.mosaalgorithm import MOSAAlgorithm
 from pynguin.ga.operators.crossover import CrossOverFunction
 from pynguin.ga.operators.ranking import RankingFunction
 from pynguin.ga.operators.selection import SelectionFunction
-from pynguin.testcase.execution import ExecutionTracer
+from pynguin.instrumentation.tracer import ExecutionTracer
 from pynguin.testcase.execution import TestCaseExecutor
