+++ tests/testcase/execution/test_executionmocking.py
@@ -12,8 +12,8 @@ import threading
 from pynguin.analyses.constants import EmptyConstantProvider
 from pynguin.analyses.module import generate_test_cluster
 from pynguin.analyses.seeding import AstToTestCaseTransformer
+from pynguin.instrumentation.tracer import ExecutionTracer
 from pynguin.testcase.execution import ExecutionContext
-from pynguin.testcase.execution import ExecutionTracer
 from pynguin.testcase.execution import ModuleProvider
 from pynguin.testcase.execution import TestCaseExecutor
 from pynguin.testcase.mocking import MockedLogger
