+++ tests/testcase/execution/test_testcaseexecutor_integration.py
@@ -22,8 +22,8 @@ from pynguin.analyses.constants import EmptyConstantProvider
 from pynguin.analyses.module import generate_test_cluster
 from pynguin.analyses.seeding import AstToTestCaseTransformer
 from pynguin.instrumentation.machinery import install_import_hook
+from pynguin.instrumentation.tracer import ExecutionTracer
 from pynguin.testcase.execution import ExecutionResult
-from pynguin.testcase.execution import ExecutionTracer
 from pynguin.testcase.execution import ModuleProvider
 from pynguin.testcase.execution import TestCaseExecutor
 from pynguin.testcase.statement import IntPrimitiveStatement
